requirements.txt
aiofiles==23.2.1
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pycparser==2.22
pycryptodome==3.20.0
sniffio==1.3.1
